Use context for auth
