+++ src/App.js
@@ -4,6 +4,7 @@ import { CssBaseline } from '@material-ui/core';
 import { Router } from '@reach/router';
 import firebase from 'firebase/app';
 
+import { AuthProvider } from 'auth';
 import * as config from 'config';
 import { Home } from 'pages';
 import theme from 'theme';
@@ -12,11 +13,13 @@ firebase.initializeApp(config.firebase);
 
 export default function App() {
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <Router>
-        <Home path="/" />
-      </Router>
-    </ThemeProvider>
+    <AuthProvider>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <Router>
+          <Home path="/" />
+        </Router>
+      </ThemeProvider>
+    </AuthProvider>
   );
 }
