+++ src/auth/index.js
@@ -1,2 +1,3 @@
+export { default as AuthProvider } from './AuthProvider';
 export { default as useAuth } from './useAuth';
 export { default as useAuthUi } from './useAuthUi';
